Make check box scores update
Hi Joseph,

Here's a working example of the check box values updating with the score. Please take a look at it, and also please take a look at my last code review to refactor some of the functions.

Let's finish this project up and move on to the next! We (or you) can always come back to it later on.

Colleen